Create questrade_api.php
A library of functions meant to simplify use of Questrade's API. The code is quite general using CURL to issue GET and POST requests against a parameter driven endpoint. POST requests may be sent URL encoded or in JSON format. For non-experimental use improvements are suggested for the parameter encoding sections.